add logs and transfer common method in trait
