+++ src/Tickets/Commerce/Order_Modifiers/Traits/Fee_Types.php
@@ -9,6 +9,7 @@ declare( strict_types=1 );
 
 namespace TEC\Tickets\Commerce\Order_Modifiers\Traits;
 
+use Exception;
 use TEC\Tickets\Commerce\Order_Modifiers\Repositories\Order_Modifiers;
 
 /**
@@ -110,5 +111,46 @@ trait Fee_Types {
 		}
 
 		return $ticket_fees;
+	}
+
+	/**
+	 * Update the fees for a ticket.
+	 *
+	 * @since TBD
+	 *
+	 * @param int   $ticket_id The ticket ID.
+	 * @param int[] $fees      The fees to update.
+	 *
+	 * @return void
+	 *
+	 * @throws Exception If the fees are not selectable.
+	 */
+	protected function update_fees_for_ticket( $ticket_id, $fees ) {
+		// Validate that the fees are actually selectable.
+		$all_fees        = $this->get_all_fees();
+		$selectable_fees = wp_list_pluck( $this->get_selectable_fees( $all_fees ), 'id', 'id' );
+		$invalid_fees    = [];
+		foreach ( $fees as $fee ) {
+			if ( ! array_key_exists( $fee, $selectable_fees ) ) {
+				$invalid_fees[] = $fee;
+			}
+		}
+
+		if ( ! empty( $invalid_fees ) ) {
+			throw new Exception(
+				sprintf(
+					/* translators: %s: The invalid fees. */
+					__( 'The following fees are not selectable: %s', 'event-tickets' ),
+					implode( ', ', $invalid_fees )
+				),
+				400
+			);
+		}
+
+		// Ensure that the fees are integers.
+		$fee_ids = array_map( 'absint', $fees );
+
+		$this->manager->delete_relationships_by_post( $ticket_id );
+		$this->manager->sync_modifier_relationships( $fee_ids, [ $ticket_id ] );
 	}
 }
